Fix: Array encoding when creating arbitrary transactions
